move Wrapper to common folder for reuse in MoviesPage and PeoplePage
